+++ lisp/eab-minimal.el
@@ -34,7 +34,7 @@
                    (prin1-to-string binding)))
       (describe-key-briefly key insert untranslated))))
 
-;; TODO: timers depend on keboard speed
+;; DONE: timers depend on keboard speed
 (defmacro eab/do-action (&rest body)
   ""
   `'(:def (lambda ()
@@ -48,8 +48,13 @@
   (let (executing-kbd-macro defining-kbd-macro)
     (run-with-timer 0.01 nil (ilam-no-def (execute-kbd-macro (read-kbd-macro "C-]"))))
     (run-with-timer 0.02 nil (ilam-no-def (execute-kbd-macro (read-kbd-macro "M-g"))
-                                  (setq kill-ring (cdr kill-ring))))
+                                          (pop kill-ring)
+                                          (setq kill-ring-yank-pointer kill-ring)))
     (recursive-edit)))
+;; TODO unsafe (setq kill-ring (cdr kill-ring))? see kill-new
+;; (pop kill-ring)
+;; (setq kill-ring-yank-pointer kill-ring)
+
 
 (defun eab/or-self-insert (command)
   (if (use-region-p)